+++ studio-server/src/main/java/com/coremedia/labs/plugins/adapters/cloudinary/CloudinaryOptions.java
@@ -1,19 +1,22 @@
 package com.coremedia.labs.plugins.adapters.cloudinary;
 
-public class CloudinaryImportOptions {
+public class CloudinaryOptions {
   private static final boolean DEFAULT_ASSET_ID_MODE_ENABLED = false;
+  private static final String DEFAULT_SEARCH_QUERY = "filename=%1$s";
   private static final boolean DEFAULT_IMPORT_VIDEO_URL = false;
   private static final boolean DEFAULT_IMPORT_VIDEO_BLOB = true;
   private final boolean assetIdModeEnabled;
   private final String importPublicIdAs;
+  private final String searchQuery;
   private final boolean importVideoURL;
   private final boolean importVideoBlob;
   private final Integer importImageMaxWidth;
   private final Integer importImageQuality;
 
-  public CloudinaryImportOptions(CloudinaryContentHubSettings settings) {
+  public CloudinaryOptions(CloudinaryContentHubSettings settings) {
     assetIdModeEnabled = getBooleanOption(settings.getAssetIdModeEnabled(), DEFAULT_ASSET_ID_MODE_ENABLED);
     importPublicIdAs = settings.getImportPublicIdAs();
+    searchQuery = getStringOption(settings.getSearchQuery(), DEFAULT_SEARCH_QUERY);
     importVideoURL = getBooleanOption(settings.getImportVideoURL(), DEFAULT_IMPORT_VIDEO_URL);
     importVideoBlob = getBooleanOption(settings.getImportVideoBlob(), DEFAULT_IMPORT_VIDEO_BLOB);
     importImageMaxWidth = settings.getImportImageMaxWidth();
@@ -26,6 +29,10 @@ public class CloudinaryImportOptions {
 
   public String getImportPublicIdAs() {
     return importPublicIdAs;
+  }
+
+  public String getSearchQuery() {
+    return searchQuery;
   }
 
   public boolean isImportVideoURL() {
@@ -50,11 +57,19 @@ public class CloudinaryImportOptions {
     return Boolean.parseBoolean(settingValue);
   }
 
+  private String getStringOption(String settingValue, String defaultValue) {
+    if(settingValue == null)
+      return defaultValue;
+    return settingValue;
+  }
+
+
   @Override
   public String toString() {
-    return "CloudinaryImportOptions{" +
+    return "CloudinaryOptions{" +
             "assetIdModeEnabled=" + assetIdModeEnabled +
             ", importPublicIdAs='" + importPublicIdAs + '\'' +
+            ", searchQuery='" + searchQuery + '\'' +
             ", importVideoURL=" + importVideoURL +
             ", importVideoBlob=" + importVideoBlob +
             ", importImageMaxWidth=" + importImageMaxWidth +